Move internal code to core module and update imports accordingly
No changes to external functionality, improvements to organize internal
modules only.
